+++ frontend/package.json
@@ -13,6 +13,8 @@
     "@types/react-syntax-highlighter": "^15.5.9",
     "axios": "^1.5.0",
     "dotenv": "^16.3.1",
+    "embla-carousel": "^8.0.2",
+    "embla-carousel-react": "^8.0.2",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-prettier": "^5.0.0",
     "framer-motion": "^10.16.11",
