+++ frontend/src/app/App.tsx
@@ -1,0 +1,18 @@
+import '../index.css';
+import { RecoilRoot } from 'recoil';
+import { GlobalStyle } from './style/globalStyle.ts';
+import { RouterProvider } from 'react-router-dom';
+import { router } from './router/index.ts';
+
+function App() {
+  return (
+    <>
+      <RecoilRoot>
+        <GlobalStyle />
+        <RouterProvider router={router} />
+      </RecoilRoot>
+    </>
+  );
+}
+
+export default App;
