+++ frontend/src/entities/filter/index.ts
@@ -1,1 +1,1 @@
-export { DropDown } from './ui/Dropdown';
+// export { DropDown } from './ui/Dropdown';
