+++ frontend/src/entities/textEdit/index.tsx
@@ -1,0 +1,12 @@
+// import * as textEdit from './index';
+// // import React from 'react';
+
+// export default function index() {
+//   return (
+//     <div className="relative flex flex-col w-screen h-screen bg-black px-[0rem]">
+//       <textEdit.headerInput />
+//       <textEdit.markdownView />
+//       <textEdit.bottomButtons />
+//     </div>
+//   );
+// }
