+++ frontend/src/index.css
@@ -6,20 +6,20 @@
 @font-face {
     font-family: 'Pre-R';
     src: local('Pre-R'),
-    url('shared/fonts/Pretendard/Pretendard-Regular.woff');
+    url('shared/assets/fonts/Pretendard/Pretendard-Regular.woff');
     font-weight: 400;
 }
 
 @font-face {
     font-family: 'Pre-S';
     src: local('Pre-S'),
-    url('shared/fonts/Pretendard/Pretendard-SemiBold.woff');
+    url('shared/assets/fonts/Pretendard/Pretendard-SemiBold.woff');
     font-weight: 400;
 }
 
 @font-face {
     font-family: 'Bayon';
     src: local('Bayon'),
-    url('shared/fonts/Bayon/Bayon-Regular.ttf');
+    url('shared/assets/fonts/Bayon/Bayon-Regular.ttf');
     font-weight: 400;
 }
