--- frontend/src/setupProxy.ts
@@ -1,13 +1,0 @@
-
-// import { createProxyMiddleware } from 'http-proxy-middleware';
-// import { Application } from 'express';
-
-// module.exports = function(app: Application) {
-//   app.use(
-//     '/members',
-//     createProxyMiddleware({
-//       target: 'http://localhost:8080',
-//       changeOrigin: true,
-//     }),
-//   );
-// };
