+++ frontend/src/shared/assets/image/Icon-Point.png
