+++ frontend/src/shared/assets/image/pointBanner.png
