+++ frontend/src/shared/types/fonts.d.ts
