+++ frontend/src/shared/ui/Footer.tsx
