+++ frontend/src/shared/ui/NavBar.tsx
