--- frontend/src/theme.ts
@@ -1,7 +1,0 @@
-import { DefaultTheme } from 'styled-components'
-
-export const theme: DefaultTheme = {
-  bgColor: '#f5f6fa',
-  textColor: '#2f3640',
-  accentColor: '#4cd137'
-}
